+++ app/src/main/kotlin/com/twain/interprep/MainActivity.kt
@@ -12,7 +12,7 @@ import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.compose.rememberNavController
 import com.twain.interprep.data.ui.QuoteData
 import com.twain.interprep.helper.PrefManager
-import com.twain.interprep.presentation.navigation.NavGraph
+import com.twain.interprep.presentation.navigation.OnboardingNavGraph
 import com.twain.interprep.presentation.ui.modules.interview.QuotesViewModel
 import com.twain.interprep.presentation.ui.theme.InterPrepTheme
 import com.twain.interprep.presentation.ui.theme.MaterialColorPalette
@@ -42,7 +42,7 @@ class MainActivity : ComponentActivity() {
                     color = MaterialColorPalette.surface
                 ) {
                     val navController = rememberNavController()
-                    NavGraph(navController = navController, prefManager)
+                    OnboardingNavGraph(navController = navController, prefManager)
                 }
             }
         }
