+++ app/src/main/res/values/strings.xml
@@ -107,6 +107,7 @@
     <string name="error_message_invalid_url">URL is not valid</string>
     <string name="error_message_invalid_number">Number is not valid</string>
 
+    <string name="header_greeting_screen">How would you like us to greet you ?</string>
     <string name="header_add_interview">Add your interview information here.</string>
     <string name="header_edit_interview">Update your scheduled interview details.</string>
     <string name="header_add_note">Add your interview notes</string>
@@ -125,6 +126,7 @@
     <string name="button_add_link">Add Link</string>
     <string name="button_delete_link">Delete Link</string>
     <string name="button_get_started">"Get Started"</string>
+    <string name="button_next">Next</string>
 
     <string name="interview_status_no_update">No Update</string>
     <string name="interview_status_next_round">Next Round</string>
